+++ changelog.md
@@ -7,6 +7,7 @@
   * Same depth scaling (up to +20%) and full moon bonus (+20%) now also apply to the chance of equipment (weapons, tools, armor) getting enchanted
   * Iron and Diamond armor are now more likely to be picked over weaker materials when underground or during a full moon
 * Enderman can no longer teleport targets farther than their follow range
+* Fixed fisher mobs no longer attacking and ignoring line of sight
 
 # 4.2.3.0
 * Pathfinding